Replaced advanced with restore
