+++ README.md
@@ -10,10 +10,13 @@ Try the demo on the gh-pages site [here!](https://3djakob.github.io/react-tinder
 
 ## The code
 
-Feel free to check out the code in `react-tinder-card-demo/src/examples` for a usage example of how this application uses the module.
+Feel free to check out the code in [react-tinder-card-demo/src/examples](https://github.com/3DJakob/react-tinder-card-demo/tree/master/src/examples) for a usage example of how this application uses the module.
 
+### Simple
 The simple example is the minimum code needed to get you started.
 
-The advanced example implements a state to dynamically remove swiped elements as well as using buttons to trigger swipes.
+### Advanced
+The advanced example implements a state with references to swipe and restore cards using buttons.
 
+## Demo
 Both code examples can be tested on the [demo page.](https://3djakob.github.io/react-tinder-card-demo/)
